Create README.md
Hello,
I'm new to the coding world.  I have a lot of experience in many creative mediums, like film making, songwriting, photography, digital art, drafting, and finally now is the time to see what web development has to offer. I'm inspired by what's being built these days, it seems to be never ending innovation. Some exciting languages and tools to learn and create with and such an amazing community of people with resources to help each other. I'm so happy to be here!